Add files via upload
Updated display.py and server.py. 
display.py is now acting as the client process.
An additional thread is created in display.py once the chat window is displayed. This thread runs an infinite loop that is meant to check for incoming packets from server. Does not appear to be working at present. Tested the ADD command message only (not working!!!).